# core-lending-credit-cards-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-credit-cards
  description: Credit Cards Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/credit-cards
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}/billing-cycles/{ccBillingCycleId}/statements/{ccStatementId}:
    get:
      tags:
        - CcStatement
      summary: Get a statement by ID
      operationId: getById_5
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccStatementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcStatementDTO'
    put:
      tags:
        - CcStatement
      summary: Update a statement record
      operationId: update_5
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccStatementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcStatementDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcStatementDTO'
    delete:
      tags:
        - CcStatement
      summary: Delete a statement record
      operationId: delete_5
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccStatementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}/billing-cycles/{ccBillingCycleId}/statements:
    get:
      tags:
        - CcStatement
      summary: List or search statements within a billing cycle
      operationId: findAll_5
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.ccBillingCycleId
          in: query
          description: Exact filter for cc billing cycle id
          required: false
          schema:
            type: string
        - name: filters.statementDate
          in: query
          description: Exact filter for statement date
          required: false
          schema:
            type: string
        - name: filters.statementBalance
          in: query
          description: Exact filter for statement balance
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[statementBalance].from
          in: query
          description: Filter statement balance from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[statementBalance].to
          in: query
          description: Filter statement balance to value
          required: false
          schema:
            type: string
        - name: filters.minimumPayment
          in: query
          description: Exact filter for minimum payment
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[minimumPayment].from
          in: query
          description: Filter minimum payment from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[minimumPayment].to
          in: query
          description: Filter minimum payment to value
          required: false
          schema:
            type: string
        - name: filters.paymentDueDate
          in: query
          description: Exact filter for payment due date
          required: false
          schema:
            type: string
        - name: filters.isPaid
          in: query
          description: Exact filter for is paid
          required: false
          schema:
            type: boolean
        - name: filters.paidDate
          in: query
          description: Exact filter for paid date
          required: false
          schema:
            type: string
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCcStatementDTO'
    post:
      tags:
        - CcStatement
      summary: Create a new statement
      operationId: create_5
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcStatementDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcStatementDTO'
components:
  schemas:
...
    CcStatementDTO:
      required:
        - ccBillingCycleId
        - isPaid
        - minimumPayment
        - paymentDueDate
        - statementBalance
        - statementDate
      type: object
      properties:
        ccStatementId:
          type: string
          format: uuid
          readOnly: true
        ccBillingCycleId:
          type: string
          format: uuid
        statementDate:
          type: string
          format: date
        statementBalance:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: -999999999.99
          exclusiveMinimum: false
          type: number
        minimumPayment:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        paymentDueDate:
          type: string
          format: date
        isPaid:
          type: boolean
        paidDate:
          type: string
          format: date
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseCcStatementDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CcStatementDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.